[Array] Refactor how the allocators are used
Allocator type is now by default AllocatorBase which allows us to pass allocators of different types without the need to create new type of an Array for every change of allocator. Allocators are passed in by pointer and are saved like that. User can still provide custom Allocator type to the Array for performance reasons.